@@ -1,22 +1,22 @@
-import React from "react";
-const styles = require("./styles.module.scss");
+import React from 'react';
+const styles = require('./styles.module.scss');
 
-interface Props {
-  src: any;
-  caption: string;
-  className?: any;
-  height?: string;
+interface IProps {
+    src: any;
+    caption: string;
+    className?: any;
+    height?: string;
 }
 
-const ImageHoverCaption = (props: Props) => {
-  return (
-    <div className={styles.img_container}>
-      <img src={props.src} height={props.height} className={`${styles.img_styles} ${props.className}`} />
-      <div className={styles.middle}>
-        <div className={styles.text}>{props.caption}</div>
-      </div>
-    </div>
-  );
+const ImageHoverCaption = (props: IProps): any => {
+    return (
+        <div className={styles.img_container}>
+            <img src={props.src} height={props.height} className={`${styles.img_styles} ${props.className}`} />
+            <div className={styles.middle}>
+                <div className={styles.text}>{props.caption}</div>
+            </div>
+        </div>
+    );
 };
 
 export default ImageHoverCaption;
