@@ -1,6 +1,5 @@
-export {SolarSystem, Moons, Planets} from "./backyard";
-export {About} from "./about";
-export {default as NotFound} from "./not-found";
-export {default as Loading} from "./loading";
-export {default as Home} from "./home";
-export {default as Alone} from "./alone";
+export {About} from './about';
+export {default as NotFound} from './not-found';
+export {default as SplashScreen} from './splash-screen';
+export {default as Home} from './home';
+export {default as Alone} from './alone';
