@@ -1,11 +1,11 @@
-// import React, {useState, useCallback} from "react";
+import React, {useState, useCallback} from 'react';
 
-// import {Parallax, ParallaxLayer} from "react-spring/renderprops-addons";
+import {Parallax, ParallaxLayer} from 'react-spring/renderprops-addons';
 
-// const LandingPage = () => {
-//   const parallax = React.useRef(null);
+const LandingPage = () => {
+    const parallax = React.useRef(null);
 
-//   return <div style={{backgroundColor: "pink"}}>TEST</div>;
-// };
+    return <div style={{backgroundColor: 'pink'}}>TEST</div>;
+};
 
 // export default LandingPage;
