@@ -1,15 +1,15 @@
-export {default as LayoutMain} from "./layout-main";
-export {default as NavBar} from "./top-nav";
-export {default as SideNav} from "./side-nav";
-export {default as Hero} from "./hero";
-export {default as ParticleComponent} from "./particles";
-export {default as HomeFeature} from "./home-feature";
-export {default as HomeFeatureRight} from "./home-feature-right";
-export {default as VerticalLine} from "./vertical-line";
-export {default as UfoMouseParallax} from "./ufo-mouse-parallax";
-export {default as Sun} from "./sun";
-// export {default as HorizontalEarthParallax} from "./horizontal-earth";
-export {default as Popup} from "./pop-up";
-export {default as InfoCard1} from "./info-card-1";
-export {default as ImgHoverCaption} from "./img-hover-caption";
-export {default as MercuryDaysCalculator} from "./mercury-days-calculator";
+export {default as LayoutMain} from './layout-main';
+export {default as NavBar} from './top-nav';
+export {default as SideNav} from './side-nav';
+export {default as Hero} from './hero';
+export {default as ParticleComponent} from './particles';
+export {default as HomeFeature} from './home-feature';
+export {default as HomeFeatureRight} from './home-feature-right';
+export {default as VerticalLine} from './atoms/vertical-line';
+export {default as UfoMouseParallax} from './atoms/ufo-mouse-parallax';
+export {default as Sun} from './atoms/sun';
+export {default as HorizontalEarthParallax} from './horizontal-earth';
+export {default as Popup} from './pop-up';
+export {default as InfoCard1} from './info-card-1';
+export {default as ImgHoverCaption} from './img-hover-caption';
+export {default as MercuryDaysCalculator} from './mercury-days-calculator';
