@@ -1,12 +1,12 @@
-import {useEffect} from "react";
-import {useLocation} from "react-router-dom";
+import {useEffect} from 'react';
+import {useLocation} from 'react-router-dom';
 
 export default function ScrollToTop() {
-  const {pathname} = useLocation();
+    const {pathname} = useLocation();
 
-  useEffect(() => {
-    window.scrollTo(0, 0);
-  }, [pathname]);
+    useEffect(() => {
+        window.scrollTo(0, 0);
+    }, [pathname]);
 
-  return null;
+    return null;
 }
