@@ -1,13 +1,13 @@
-const webpack = require("webpack");
-const merge = require("webpack-merge");
-const parts = require("./webpack.parts");
+const webpack = require('webpack');
+const merge = require('webpack-merge');
+const parts = require('./webpack.parts');
 
 developmentConfig = () =>
     merge([
         {
             plugins: [new webpack.HotModuleReplacementPlugin()],
         },
-        parts.generateSourceMaps({type: "eval-source-map"}),
+        parts.generateSourceMaps({type: 'eval-source-map'}),
         parts.devServer({
             host: process.env.HOST,
             port: process.env.PORT,
